Docs and more.  Now dial() and listen() return something.
Socket.dial() returns a Dialer, which is nice since it's not always
convenient to sift through the Socket.dialers list.  Same thing with
Socket.listen(), but for a Listener naturally.  Also removed the method
Socket.new_contexts() because I hated it.  (Socket.new_context() is
still around though, because I love it.)

Also the polyamorous flag gets set on Pair1 before listen/dial now, so
it's nice to get rid of that race condition.

Also lots of docs improvements.

This commit is way less granular than I like to do, but on the other
hand I'm pretty tired today, and it would be work to get a nice git
history.