Fix race condition in test.
test_sub_sock_options was relying on a race condition, where the
background thread would receive the first message before the second was
actually sent.  This happened because I incorrectly believed the message
buffer depth was greater than 1 by default, so I did not realize it was
a race condition.