Initial stab of adding wheel building with GitHub Actions.
This replaces Azure Pipelines, which was somewhat unwieldy.  There are
still some issues with the reliability of the tests, but the CI itself
is set up OK now.

Still to come: a GitHub Action that only tests a single configuration on
Linux, because that is more helpful to quickly see if PRs are working
instead of waiting for Windows, Linux, and Mac wheels to build.

Maybe future work: seeing if we can get wheels built on arm32.  Wouldn't
that be something?