Make tests faster and remove some hardcoded sleeps.
The sleeps are not reliable; they're usually too long for local
development, and sometimes too short for CI.  This brings pain.  The
sleeps are *mostly* just waiting for connections to happen in the
background, so we can just check that directly for the most part.  Also
shrank some timeouts where it seemed possible.