Fix racy, buggy test.
We were testing for number of connected pipes, which was terrifically
fragile.  The underlying socket immediately tries to reconnect, and
sometimes does reconnect before we actually get our assertions in.  We
have replaced the buggy, racy tests with one that just checks that the
post_remove callback is called; this ensures the pipe has been closed.