No sys.path mangling.
Not sure if this was ever necessary, but now it is causing failures on
readthedocs; rtd installs the project from a wheel, so it gets installed
to site-packages. When the pynng directory gets added to sys.path, it
gets picked up before the version in site packages. This is a problem
because rtd does not build the package "in place", so there are no
shared libraries where they are expected.

At least I *hope* that is what is going on.