Get v3.5.1 to build in docker image.
Seems like the root cause is that when mbedtls moved to use
GnuInstallDirs, the install directory for the library changed. Which
makes sense. I couldn't figure out how to get mbedtls to install do lib,
so I just check if lib64 exists, and if so move to lib.

Doesn't feel good, but there it is.