Fix CI build failure.
moving the lib64 to lib unconditionally causes the build to fail when
the 2nd wheel is getting built, because then the destination directory
exists. We also moved to copying instead of moving, because that will
prevent the mbedtls lib from doing extra work when it finds out that its
targets are missing.