Fix i686 Linux build failures.
mbedtls fails to build on 32-bit Linux without passing some compiler
options. We pass them through an environment variable, which is not the
best way to do it, but here we are.

Hopefully this will be able to be removed whenever mbedtls pushes a new
release; seems that this has been fixed upstream in merge
https://github.com/Mbed-TLS/mbedtls/pull/8406.